+++ dependencyLibs/Mavlink/src/ardupilotmega/msg_gopro_set_response.java
@@ -141,8 +141,8 @@ public class msg_gopro_set_response extends MAVLinkMessage {
     public JSONObject toJSON() throws JSONException {
         final JSONObject jo = getJSONheader();
         
-        jo.put("cmd_id", cmd_id);
-        jo.put("status", status);
+        jo.put("cmd_id", (int)cmd_id);
+        jo.put("status", (int)status);
         
         
         return jo;
