+++ dependencyLibs/Mavlink/src/ardupilotmega/msg_rally_point.java
@@ -221,16 +221,16 @@ public class msg_rally_point extends MAVLinkMessage {
     public JSONObject toJSON() throws JSONException {
         final JSONObject jo = getJSONheader();
         
-        jo.put("lat", lat);
-        jo.put("lng", lng);
-        jo.put("alt", alt);
-        jo.put("break_alt", break_alt);
-        jo.put("land_dir", land_dir);
-        jo.put("target_system", target_system);
-        jo.put("target_component", target_component);
-        jo.put("idx", idx);
-        jo.put("count", count);
-        jo.put("flags", flags);
+        jo.put("lat", (int)lat);
+        jo.put("lng", (int)lng);
+        jo.put("alt", (int)alt);
+        jo.put("break_alt", (int)break_alt);
+        jo.put("land_dir", (int)land_dir);
+        jo.put("target_system", (int)target_system);
+        jo.put("target_component", (int)target_component);
+        jo.put("idx", (int)idx);
+        jo.put("count", (int)count);
+        jo.put("flags", (int)flags);
         
         
         return jo;
