+++ dependencyLibs/Mavlink/src/ardupilotmega/msg_remote_log_block_status.java
@@ -161,10 +161,10 @@ public class msg_remote_log_block_status extends MAVLinkMessage {
     public JSONObject toJSON() throws JSONException {
         final JSONObject jo = getJSONheader();
         
-        jo.put("seqno", seqno);
-        jo.put("target_system", target_system);
-        jo.put("target_component", target_component);
-        jo.put("status", status);
+        jo.put("seqno", (long)seqno);
+        jo.put("target_system", (int)target_system);
+        jo.put("target_component", (int)target_component);
+        jo.put("status", (int)status);
         
         
         return jo;
